# Conflicts:
#	clients/src/main/resources/i18n/message.properties
#	customer-service/src/main/java/com/unicoin/customer/form/AddCustomerForm.java

+++ customer-service/src/main/java/com/unicoin/customer/repository/RoleRepository.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     List<Role> findAllByStatus(boolean status);
+
+
 }
